requirements.txt
django
django-admin-bootstrapped
django-bootstrap3
django-lazysignup
mysqlclient
